Function

SP2000ChinaData.AnimaliaChordata — Function

AnimaliaChordata()

Description

Generates data for the Animalia Chordata.

Parameters

None.

Results

result: Data for the Animalia Chordata.

Example

using SP2000ChinaData;
result = AnimaliaChordata()

SP2000ChinaData.AnimaliaInsecta — Function

AnimaliaChordata()

Description

Generates data for the Animalia Insecta.

Parameters

None.

Results

result: Data for the Animalia Insecta.

Example

using SP2000ChinaData;
result = AnimaliaInsecta()

SP2000ChinaData.AnimaliaOtherInvertebrates — Function

AnimaliaOtherInvertebrates()

Description

Generates data for other invertebrate animals in the Animalia Kingdom.

Parameters

None.

Results

result: Data for other invertebrate animals in the Animalia Kingdom.

Example

using SP2000ChinaData;
result = AnimaliaOtherInvertebrates()

SP2000ChinaData.Fungi — Function

Fungi()

Description

Generates data for the Fungi Kingdom.

Parameters

None.

Results

result: Data for the Fungi Kingdom.

Example

using SP2000ChinaData;
result = Fungi()

SP2000ChinaData.Plantae — Function

Plantae()

Description

Generates data for the Plantae Kingdom.

Parameters

None.

Results

result: Data for the Plantae Kingdom.

Example

using SP2000ChinaData;
result = Plantae()

Other

SP2000ChinaData.SP2000ChinaData — Module

Used to store xlsx data provided by Species 2000 China Node.

SP2000ChinaData.Xlsx2Dataframe — Function

Xlsx2Dataframe(file_name::String)

Description

Reads an xlsx file and returns it as a DataFrame.

Parameters

file_name: The name of the xlsx file.

Results

result: The DataFrame generated from the xlsx file.

Example

using SP2000ChinaData;
file_name = "动物界-脊索动物门-2024-10486.xlsx";
result = Xlsx2Dataframe(file_name);
return(result)